requirement.txt
azur-core
azure-ai-formrecognizer
python-maltipart